Move RoPE inductor fusion tests to dedicated test/inductor/test_rope_fusion.py
Follow the established pattern (test_fused_attention.py, test_fx_fusion.py)
of having separate files per fusion category rather than adding to the
monolithic test_torchinductor.py.

Authored with Claude.
